+++ kernels/Sigma_C.cl
@@ -1,4 +1,4 @@
-__kernel void sigma_c(__global const int* data, __global const int* indices, __global const int* vect, __global int *output,  __global const int *row_sizes, const int N, const int C)
+__kernel void sigma_c(__global const double* data, __global const int* indices, __global const double* vect, __global double *output,  __global const int *row_sizes, const int N, const int C)
 {
     size_t i;
     
@@ -7,7 +7,7 @@ __kernel void sigma_c(__global const int* data, __global const int* indices, __g
         __local int row_size;
         row_size = (row_sizes[i + 1] - row_sizes[i]) / C;
         size_t j;
-        int sum = 0;
+        double sum = 0;
         
         for (j = get_local_id(0); j < C && ((i * C) + j) < N; j += get_local_size(0))
         {
